refactor(rt): split SoC modules into separate sub-modules
... to make the code structure clear.
